+++ python/src/main/java/net/maswag/falcaun/python/PythonNumericSUL.java
@@ -1,10 +1,14 @@
-package net.maswag.falcaun;
+package net.maswag.falcaun.python;
 
 import de.learnlib.exception.SULException;
 import de.learnlib.sul.SUL;
 import jep.JepException;
 import lombok.Getter;
 import net.automatalib.word.Word;
+import net.maswag.falcaun.IODiscreteSignal;
+import net.maswag.falcaun.IOSignal;
+import net.maswag.falcaun.IOSignalPiece;
+import net.maswag.falcaun.NumericSUL;
 
 import javax.annotation.Nonnull;
 import javax.annotation.Nullable;
