+++ python/src/test/java/net/maswag/falcaun/python/PythonNumericSULTest.java
@@ -1,6 +1,11 @@
-package net.maswag.falcaun;
+package net.maswag.falcaun.python;
 
 import org.junit.jupiter.api.Test;
+
+import net.maswag.falcaun.IOSignal;
+import net.maswag.falcaun.IOSignalPiece;
+import net.maswag.falcaun.python.PythonNumericSUL;
+
 import java.util.Arrays;
 
 import static org.junit.jupiter.api.Assertions.assertFalse;
